+++ packages/adminui-vue/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
  - `UTableEntity` filters for different columns no more overrides each other
+ - FTS navbar widget will not intercept `Ctrl+F` in case fts is disabled in config and widget is hidden
 
 ## [1.9.25] - 2020-02-23
 ### Removed