+++ packages/adminui-vue/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+ - error while opening a multilanguage attribute editor (`truncTimeToUTCNull` will skip such attributes)
  - `UTableEntity` filters for different columns no more overrides each other
  - FTS navbar widget will not intercept `Ctrl+F` in case fts is disabled in config and widget is hidden
  - `UAutoField` remove mistaken added properties for `Text` field, added missing input handler